requirements.txt
asgiref>=3.8,<4.0
Django>=5.1,<6.0
django-environ>=0.12,<1.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.7,<5.0
mysqlclient>=2.2,<3.0
sqlparse>=0.5,<1.0
tzdata>=2025.2,<2026.0
gunicorn>=21.2,<22.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.4,<5.0
